// web/src/generated/models/reviewBatchDocumentCodingRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Priv-View Core API
 * Phase-one Core API for authentication, hierarchical tenants, clients, matters, and matter metadata definitions.
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewBatchDocumentCodingReadReviewStatus } from './reviewBatchDocumentCodingReadReviewStatus';
import type { ReviewBatchReviewerValueRead } from './reviewBatchReviewerValueRead';
import type { ReviewBatchRunValueRead } from './reviewBatchRunValueRead';

export interface ReviewBatchDocumentCodingRead {
  matter_document_id: string;
  review_status: ReviewBatchDocumentCodingReadReviewStatus;
  values: ReviewBatchRunValueRead[];
  reviewer_values: ReviewBatchReviewerValueRead[];
}